move the tool to another path
